# lib/api-zod/src/generated/types/providerKey.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EKYVA AI gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderKeyStatus } from './providerKeyStatus';

export interface ProviderKey {
  provider: string;
  configured_at: Date;
  status: ProviderKeyStatus;
}